refactor(api): move target profile repository to its bounded context
